+++ HISTORY.md
@@ -2,6 +2,7 @@
 * added LICENSE
 * added HISTORY
 * setting logger level before first append used incorrect level (fixed)
+* added "fine" logger level
 
 ### v3.1 2019-08-25
 * conversion of 3.0 version